cody tidy, re-org + removing external js dependencies
