+++ src/main/java/com/redhat/sso/ninja/InitServlet.java
@@ -2,19 +2,12 @@ package com.redhat.sso.ninja;
 
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.attribute.PosixFilePermission;
-import java.util.HashSet;
-import java.util.Set;
 
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 
-import org.apache.camel.CamelContext;
-import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.impl.DefaultCamelContext;
 import org.apache.commons.io.IOUtils;
 
 import com.redhat.sso.ninja.utils.DownloadFile;
@@ -24,7 +17,7 @@ public class InitServlet extends HttpServlet {
   @Override
   public void init(ServletConfig config) throws ServletException {
     super.init(config);
-//    Summarizer.start();
+    
     String intervalString=(String)Config.get().getOptions().get("heartbeat.interval");
     if (null==intervalString) intervalString="60000";
     int interval=Integer.parseInt(intervalString);
@@ -33,7 +26,6 @@ public class InitServlet extends HttpServlet {
     System.out.println("Heartbeat:");
     System.out.println("  Disabled: "+heartbeatDisabled);
     System.out.println("  Interval: "+interval +" (seconds)");
-    
     
     System.out.println("Initialise some dependencies (hopefully will be able to replace these someday):");
     if (!new File(GoogleDrive2.getDefaultExecutable()).exists()){
@@ -54,22 +46,22 @@ public class InitServlet extends HttpServlet {
       }
     }
     
-    System.out.println("Initalise the scripts (again, hopefully remove the need for this someday)");
-    try{
-      // copy the default scripts over to where they can be executed
-      IOUtils.copy(Config.class.getClassLoader().getResourceAsStream("scripts/github-stats.py"), new FileOutputStream("github-stats.py"));
-      IOUtils.copy(Config.class.getClassLoader().getResourceAsStream("scripts/trello-stats.py"), new FileOutputStream("trello-stats.py"));
-      
-      Set<PosixFilePermission> perms = new HashSet<PosixFilePermission>();
-      perms.add(PosixFilePermission.OWNER_READ);
-      perms.add(PosixFilePermission.OWNER_WRITE);
-      perms.add(PosixFilePermission.OWNER_EXECUTE);
-      Files.setPosixFilePermissions(new File("github-stats.py").toPath(), perms);
-      Files.setPosixFilePermissions(new File("trello-stats.py").toPath(), perms);
-      
-    }catch(IOException e){
-      
-    }
+//    System.out.println("Initalise the scripts (again, hopefully remove the need for this someday)");
+//    try{
+//      // copy the default scripts over to where they can be executed
+//      IOUtils.copy(Config.class.getClassLoader().getResourceAsStream("scripts/github-stats.py"), new FileOutputStream("github-stats.py"));
+//      IOUtils.copy(Config.class.getClassLoader().getResourceAsStream("scripts/trello-stats.py"), new FileOutputStream("trello-stats.py"));
+//      
+//      Set<PosixFilePermission> perms = new HashSet<PosixFilePermission>();
+//      perms.add(PosixFilePermission.OWNER_READ);
+//      perms.add(PosixFilePermission.OWNER_WRITE);
+//      perms.add(PosixFilePermission.OWNER_EXECUTE);
+//      Files.setPosixFilePermissions(new File("github-stats.py").toPath(), perms);
+//      Files.setPosixFilePermissions(new File("trello-stats.py").toPath(), perms);
+//      
+//    }catch(IOException e){
+//      
+//    }
 
     
     
@@ -93,7 +85,6 @@ public class InitServlet extends HttpServlet {
   @Override
   public void destroy() {
     super.destroy();
-//    Summarizer.stop();
     Heartbeat2.stop();
   }
